% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon2map-package.R
\docType{package}
\name{latlon2map-package}
\alias{latlon2map}
\alias{latlon2map-package}
\title{latlon2map: Facilitates matching lat/lon data with administrative units and other geographic shapes}
\description{
Facilitates matching lat/lon data with administrative units and other geographic shapes
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giocomai.github.io/latlon2map/}
}

}
\author{
\strong{Maintainer}: Giorgio Comai \email{g@giorgiocomai.eu}

}
\keyword{internal}
