% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{ll_app}
\alias{ll_app}
\title{Run the Shiny Application}
\usage{
ll_app(max_file_size = 100, ll_folder_path = NULL, ...)
}
\arguments{
\item{max_file_size}{Maximum file size to accept for upload expressed in MB, defaults to 100.}

\item{ll_folder_path}{If given, sets the folder to use for caching, corresponds to \code{ll_set_folder()}. Useful e.g. for Docker deployments. Defaults to NULL.}

\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
