% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_create_folders.R
\name{ll_create_folders}
\alias{ll_create_folders}
\title{Create folders to store geographic data}
\usage{
ll_create_folders(
  geo,
  level,
  resolution,
  year,
  file_type = c("shp", "zip", "rds")
)
}
\arguments{
\item{geo}{The geographic unit of reference as a two-letter code}

\item{level}{E.g. NUTS0, NUTS1, or county, state, ecc.}

\item{resolution}{Either resolution level as given by the data distributor, or generic such as "high", "low", or "default}

\item{file_type}{By defaults, it creates folder for zip, shp, and rds files.}
}
\value{

}
\description{
Create folders to store geographic data
}
