% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_export_sf_to_kml.R
\name{ll_export_sf_to_kml}
\alias{ll_export_sf_to_kml}
\title{Export sf objects into kml file that can be used with Google Earth, Google Maps, etc.}
\usage{
ll_export_sf_to_kml(
  sf,
  path,
  name = NULL,
  description = NULL,
  label_text = NULL,
  label_font = "Roboto Sans, Noto Sans, Helvetica",
  label_size = "24pt",
  label_placement = "m",
  line_colour = "000000ff",
  line_width = "3px",
  icon_url = "",
  icon_colour = "000000ff",
  fill_colour = NULL
)
}
\arguments{
\item{sf}{An object of class \code{sf}}

\item{path}{Path where to save the .kml output.}

\item{name}{Column to be used for names.}

\item{description}{Column to be used for description.}

\item{label_text}{Column to be used as label text. Defaults to NULL. Corresponds to "LABEL" element in OGR.}

\item{label_font}{Font family to be used for the font. Defaults to "Roboto Sans, Noto Sans, Helvetica"}

\item{label_size}{Size of the label. Defaults to "24pt"}

\item{label_placement}{Defaults to "m" (centre and middle-aligned). For more options, check: https://gdal.org/user/ogr_feature_style.html}

\item{line_colour}{Defaults to "#ffffffff" (i.e. white, with 100\% opacity). Line corresponds to "PEN" in OGR. Accepts 8-digit hex codes to include transparency.}

\item{line_width}{Defaults to "3pt". Line corresponds to "PEN" in OGR. Besides pt (points), other acceptable units are \code{g}: Map Ground Units (whatever the map coordinate units are), \code{px} Pixels, \code{pt} Points (1/72 inch), \code{mm} Millimeters, \code{cm} Centimeters, \verb{in} Inches.}

\item{icon_url}{Defaults to "" for no URL. Corresponds to "SYMBOL" in OGR. In case of wrong inputs, Google Earth may show you an ugly yellow pushpin instead (i.e. default to http://maps.google.com/mapfiles/kml/pushpin/ylw-pushpin.png). Available icons offered by Google available at this link: http://kml4earth.appspot.com/icons.html}

\item{icon_colour}{Defaults to "#000000ff" (i.e. black, with 100\% opacity).}

\item{fill_colour}{Defaults to NULL. Fill corresponds to "BRUSH" in OGR. If given, colour to be used for filling polygons.}
}
\value{

}
\description{
Attention: this function requires libkml.
}
\details{
For further details on the exact meaning of each of the parameters, please consult the documentation of OGR (used by GDAL to pass parameters to .kml): https://gdal.org/user/ogr_feature_style.html
}
