% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_find_file.R
\name{ll_find_file}
\alias{ll_find_file}
\title{Find file names. Mostly used internally}
\usage{
ll_find_file(geo, level, resolution, year, name = "abl", file_type = "rds")
}
\arguments{
\item{geo}{}

\item{level}{}

\item{resolution}{}

\item{year}{}

\item{name}{Name of specific dataset being downloaded. Defaults to abl, i.e. administrative boundary line}

\item{file_type}{}
}
\value{

}
\description{
Find file names. Mostly used internally
}
