% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_lau_pt.R
\name{ll_get_lau_pt}
\alias{ll_get_lau_pt}
\title{Regions and provinces in Italy (high detail, CC-BY Istat)}
\usage{
ll_get_lau_pt(
  id = NULL,
  name = NULL,
  year = 2017,
  level = "freguesia",
  silent = FALSE
)
}
\arguments{
\item{id}{A character vector composed of six digits. Corresponds to "dicofre".}

\item{year}{Defaults to 2017 (latest and currently only available).}

\item{level}{Defaults to "freguesia". Valid value include "freguesia", "concelho", "distrito", "des_simpli".}
}
\description{
Source: https://dados.gov.pt/pt/datasets/freguesias-de-portugal/
}
\examples{
ll_set_folder(fs::path(fs::path_home_r(), "R"))
ll_get_lau_pt()
ll_get_lau_pt(name = "Porto")
}
