% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_us.R
\name{ll_get_nuts_us}
\alias{ll_get_nuts_us}
\title{Get US counties}
\usage{
ll_get_nuts_us(level = "county", resolution = "500k", year = 2018)
}
\arguments{
\item{level}{Defaults to "county". Available options are: "cd116" (for congressional districts of the 116th Congress)}

\item{resolution}{Defaults to "500k", max available resolution. Available options are: "5m" and "20m"}

\item{year}{Defaults to 2018}
}
\value{

}
\description{
Source: https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html
}
\examples{
ll_get_nuts_us(level = "county", resolution = "500k", year = 2018)
}
