% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_population_grid_high_resolution.R
\name{ll_get_population_grid_hr}
\alias{ll_get_population_grid_hr}
\title{Get High Resolution Population Density Maps + Demographic Estimates}
\usage{
ll_get_population_grid_hr(
  geo,
  match_sf = NULL,
  match_name = NULL,
  join = sf::st_intersects,
  source_url = NULL,
  silent = FALSE
)
}
\arguments{
\item{geo}{A twe letter country code, such as "IT" for Italy and "DE" for Germany}

\item{match_sf}{An sf object to me matched with the population grid. If not given, full grid is returned.}

\item{match_name}{A name to be used for local caching. It is the responsibility of the user to keept it consistent. If not given, data are not cached locally.}

\item{join}{}

\item{source_url}{A direct link to the zipped version of the csv file in the original database, if automatic download with the country code does not work. For example, for Italy this would be "https://data.humdata.org/dataset/0eb77b21-06be-42c8-9245-2edaff79952f/resource/1e96f272-7d86-4108-b4ca-5a951a8b11a0/download/population_ita_2019-07-01.csv.zip"}

\item{silent}{}
}
\value{

}
\description{
Source: https://data.humdata.org/organization/facebook
Details on methodology: https://dataforgood.fb.com/docs/methodology-high-resolution-population-density-maps-demographic-estimates/
}
