% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_world.R
\name{ll_get_world}
\alias{ll_get_world}
\title{Get countries as an sf object}
\usage{
ll_get_world(resolution = "60", year = 2016, name = NULL)
}
\arguments{
\item{resolution}{Defaults to "60", for 1:60 Million. Available values: are 20, 10, 3, 1 (1 is highest quality available)-}

\item{year}{Defaults to 2016. Available values: 2016, 2013, 2010, 2006, 2001}
}
\value{

}
\description{
Get countries as an sf object
}
