% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_match_geo.R
\name{ll_match_geo}
\alias{ll_match_geo}
\title{Matches a data frame with longitude and latitude to an sf object}
\usage{
ll_match_geo(
  data,
  longitude = 1,
  latitude = 2,
  join = sf::st_intersects,
  sample = NULL,
  match = longlat2map::ll_get_world()
)
}
\arguments{
\item{data}{A data frame or tibble with a column for longitude and one for latitude or an onject of the sf class. If an sf object is given, the longitude and latitude parameters are ignored.}

\item{longitude}{The exact column name or the column index (e.g. 1 if first column) for longitude. Defaults to 1.}

\item{latitude}{The exact column name or the column index (e.g. 1 if first column) for latitude. Defaults to 2.}

\item{join}{A function of the sf class determining the type of join. Defaults to \code{sf::st_intersects}. Check \code{?sf::st_join} for alternatives.}

\item{sample}{Defaults to NULL. If given, it runs the matching with only a subset of the original dataframe. Suggested for testing in particular when working with big datasets.}

\item{match}{An sf object to be matched with the given dataframe, defaults to \code{longlat2map::ll_get_world()}. This package facilitate obtaining alternative reference maps with functions such as \code{longlat2map::ll_get_nuts_eu()} and \code{longlat2map::ll_get_nuts_us()}}
}
\value{
An sf object with CRS 4326.
}
\description{
Matches a data frame with longitude and latitude to an sf object
}
