% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_osm_download.R
\name{ll_osm_download_it}
\alias{ll_osm_download_it}
\title{Download OSM data in geopackage format for regions, provinces, and municipalities in Italy.}
\usage{
ll_osm_download_it(
  level = "comuni",
  name = NULL,
  code = NULL,
  overwrite = FALSE,
  wget = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{level}{One of "regioni", "provincie", "comuni". Defaults to "comuni".}

\item{name}{Name of geographic entity. Check \code{ll_osm_it_gpkg} or \code{ll_get_nuts_it()} for valid names.}

\item{code}{Used in alternative to name. Check \code{ll_osm_it_gpkg} or \code{ll_get_nuts_it()} for valid values.}

\item{wget}{Logical, defaults to FALSE. If TRUE, it downloads files with wget (if available), otherwise uses default method. Setting wget to TRUE may contribute to prevent download timeouts; notice that apparent freeze of the download progress in the console are common, and mostly the download is just continuing in the background (for reference, check file size in folder.)}

\item{quiet}{Logical, defaults to FALSE. If TRUE no messages about download advancement are printed.}
}
\value{
Used only for its side effects (downloads osm data).
}
\description{
See \code{ll_osm_it_gpkg} for all available files.
}
\examples{

\dontrun{
ll_osm_download_it(level = "comuni", name = "Trento")
}
}
