% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_osm_download.R
\name{ll_osm_extract_it}
\alias{ll_osm_extract_it}
\title{Extract OSM data for regions, provinces, and municipalities in Italy.}
\usage{
ll_osm_extract_it(
  level = "comuni",
  name = NULL,
  code = NULL,
  layer = "lines",
  quiet = FALSE
)
}
\arguments{
\item{level}{One of "regioni", "provincie", "comuni". Defaults to "comuni".}

\item{name}{Name of geographic entity. Check \code{ll_osm_it_gpkg} or \code{ll_get_nuts_it()} for valid names.}

\item{code}{Used in alternative to name. Check \code{ll_osm_it_gpkg} or \code{ll_get_nuts_it()} for valid values.}

\item{layer}{Defaults to "lines". Must be one of "points", "lines", "multilinestrings", "multipolygons", or "other_relations"}

\item{quiet}{Logical, defaults to FALSE. If TRUE, supresses messages generated when reading the geopackage file.}
}
\value{
An sf object.
}
\description{
See \code{ll_osm_it_gpkg} for all available files.
}
\examples{
\dontrun{
ll_osm_extract_it(level = "comuni", name = "Trento")
}

}
