% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_lau_eu.R
\name{ll_osm_get_nuts_streets}
\alias{ll_osm_get_nuts_streets}
\title{Get all streets available in OpenStreetMap located in given NUTS.}
\usage{
ll_osm_get_nuts_streets(
  nuts_id,
  level = 3,
  resolution = 1,
  country = NULL,
  unnamed_streets = TRUE,
  nuts_boundary_sf = NULL,
  streets_sf = NULL,
  country_code_type = "eurostat",
  year = 2021
)
}
\arguments{
\item{nuts_id}{NUTS region identifier.}

\item{country}{Name of country as included in Geofabrik's datasets, does not
always match common country names or geography. For details on available
country names see the dataset included in this package: \code{ll_osm_countries}}

\item{unnamed_streets}{Defaults to \code{TRUE}. If \code{FALSE}, it drops all streets
with missing "name" or missing \code{fclass}.}

\item{streets_sf}{Defaults to \code{NULL}. If given, used to speed up processing.
Must be an \code{sf} object such as the ones output by \code{\link[=ll_osm_get_roads]{ll_osm_get_roads()}}.}

\item{country_code_type}{Defaults to "eurostat". An alternative common value
is "iso2c". See \code{countrycode::codelist} for a list of available codes.}

\item{year}{Year of LAU boundaries, defaults to most recent (2021), passed to
\code{ll_get_lau_eu()}. Available starting with 2011.}
}
\value{
An \code{sf} objects with all streets of a given NUTS regions based on
OpenStreetMap
}
\description{
Relies on the output of \code{ll_get_nuts_eu()} for the boundaries of NUTS.
}
\examples{
\dontrun{
ll_osm_get_nuts_streets(nuts_id = "PT16D", country = "portugal")
}
}
