% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_osm.R
\name{ll_osm_get_roads}
\alias{ll_osm_get_roads}
\title{Extract shape files of roads from previously downloaded}
\usage{
ll_osm_get_roads(country, silent = FALSE)
}
\arguments{
\item{country}{The name of one or more geographic entities from files
typically previously downloaded with \code{\link[=ll_osm_download]{ll_osm_download()}}.}

\item{silent}{Defaults to \code{FALSE}. If \code{TRUE}, hides copyright notice. Useful
e.g. when using this in reports or in loops. The copyright notice must
still be shown where the final output is used.}
}
\value{
All roads in a country by OpenStreetMap.
}
\description{
Extract shape files of roads from previously downloaded
}
\examples{
\dontrun{
ll_osm_get_roads(country = "Romania")
}

}
