% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ll_osm_it_gpkg}
\alias{ll_osm_it_gpkg}
\title{Geographic entities in Italy for which geopackage files are availabile}
\format{
A list of tibbles
}
\source{
\url{https://osmit-estratti.wmcloud.org/}
}
\usage{
ll_osm_it_gpkg
}
\description{
A dataset with all names of geographic entities available for direct download as geopackage files
They are used to download files with \code{ll_osm_download_it()}
}
\keyword{datasets}
