% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_lau_eu.R
\name{ll_osm_lau_streets}
\alias{ll_osm_lau_streets}
\title{Get all streets available in OpenStreetMap located in given local administrative unit.}
\usage{
ll_osm_lau_streets(gisco_id, unnamed_streets = TRUE, year = 2020)
}
\arguments{
\item{gisco_id}{Gisco identifier.}

\item{unnamed_streets}{Defaults to TRUE. If FALSE, it drops all streets with missing "name" or missing "fclass".}

\item{year}{Year of mapping, defaults to most recent (2020). Available starting with 2011.}
}
\value{

}
\description{
Relies on the output of \code{ll_get_lau_eu()} for the boundaries of local administrative units.
}
\examples{

\dontrun{
ll_osm_lau_streets(gisco_id  = "IT_022205", unnamed_streets = FALSE)
}
}
