% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_set_folder.R
\name{ll_set_folder}
\alias{ll_set_folder}
\title{Set folder for caching data}
\usage{
ll_set_folder(path = NULL)
}
\arguments{
\item{path}{A path to a location. If the folder does not exist, it will be created.}
}
\value{
The path to the caching folder, if previously set.
}
\description{
Set folder for caching data
}
