% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_grid_hr_metadata}
\alias{population_grid_hr_metadata}
\title{A data frame with links to High Resolution Population Density Maps distributed by Facebook on HDX}
\format{
A tibble
\describe{
\item{country}{Name of the country in English}
\item{country_code}{Two letter code as used by eurostat, see also \code{countrycode::codelist$eurostat}}
\item{download_ulr}{Link to zipped dataset}
\item{url}{Link to page describing the dataset}
}
}
\source{
\url{https://data.humdata.org/}
}
\usage{
population_grid_hr_metadata
}
\description{
It is used to download files with \code{ll_get_population_grid_hr()}
}
\keyword{datasets}
