% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_find_pop_centre.R
\name{ll_find_pop_centre}
\alias{ll_find_pop_centre}
\title{Find the population-weighted centre of a municipality}
\usage{
ll_find_pop_centre(
  sf_location,
  sf_population_grid,
  power = 2,
  join = sf::st_intersects
)
}
\arguments{
\item{sf_location}{}

\item{sf_population_grid}{}

\item{power}{Defaults to 2. To give more weight to cells with higher population density, raise the number of residents by the power of.}

\item{join}{Defaults to sf::st_intersects.}
}
\value{

}
\description{
Find the population-weighted centre of a municipality
}
\examples{

ll_set_folder("~/R/")
name = "Pinzolo"
sf_location <- ll_get_nuts_it(name = name, level = "lau", resolution = "high")

lau_grid_name_temp <- stringr::str_c(name, "_lau_high-st_intersects")

sf_location_grid <- ll_get_population_grid(match_sf = sf_location,
                                          match_name = lau_grid_name_temp,
                                          match_country = "IT",
                                          join = sf::st_intersects)


pop_centre <- ll_find_pop_centre(sf_location = sf_location,
                                 sf_population_grid = sf_location_grid,
                                 power = 2)
}
