% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_electoral_districts_it.R
\name{ll_get_electoral_districts_it}
\alias{ll_get_electoral_districts_it}
\title{Get Italian electoral districts (CC-BY Istat)}
\usage{
ll_get_electoral_districts_it(
  name = NULL,
  level = "Circoscrizioni_Camera",
  year = 2022,
  silent = FALSE,
  no_check_certificate = FALSE
)
}
\arguments{
\item{level}{Defaults to "Circoscrizioni_Camera". Valid values:
\itemize{
\item{"Circoscrizioni_Camera"}: Basi geografiche delle circoscrizioni elettorali - Camera dei deputati
\item{"Regioni_Senato"}: Basi geografiche delle circoscrizioni elettorali - Senato della Repubblica
\item{"CAMERA_CollegiPLURINOMINALI_2020"}: Basi geografiche dei collegi elettorali plurinominali - Camera dei deputati
\item{"CAMERA_CollegiUNINOMINALI_2020"}: Basi geografiche dei collegi elettorali uninominali - Camera dei deputati
\item{"SENATO_CollegiPLURINOMINALI_2020"}: Basi geografiche dei collegi elettorali plurinominali - Senato della Repubblica
\item{"SENATO_CollegiUNINOMINALI_2020"}: Basi geografiche dei collegi elettorali uninominali - Senato della Repubblica
\item{"UT_Collegi2020"}: Basi geografiche delle unità territoriali che formano i collegi elettorali (comuni e aree sub-comunali, limitatamente ai comuni di Torino, Genova, Milano, Roma, Napoli e Palermo con territorio ripsrtito su più di un collegio). Geografia comunale vigente alla data della pubblicazione
}}

\item{year}{Defaults to 2022 (latest available). Currently no other year accepted.}

\item{no_check_certificate}{Logical, defaults to TRUE. Enable only if certificate issues, and if you are aware of the security implications.}
}
\description{
2022 / WGS 84 / UTM zone 32N
}
\details{
Column names metadata:

\itemize{
\item{COD_REG	Codice della regione/circoscrizione elettorale del Senato della Repubblica}
\item{DEN_REG	Denominazione della regione amministrativa/circoscrizione elettorale Senato della Repubblica}
\item{COD_PRO	Codice della provincia}
\item{DEN_P_CM	Denominazione della provincia o città metropolitana}
\item{COD_CM	Codice della città metropolitana}
\item{PRO_COM	Codice del comune}
\item{DEN_COM	Denominazione del comune}
\item{CAP_DEN	Denominazione del capoluogo di provincia o città metropolitana}
\item{POP_2011	Popolazione - Censimento 2011 }
\item{ASC_COD	Codice concatenato comune e area sub-comunale}
\item{ASC_COD1	Codice progressivo area sub-comunale}
\item{ASC_COD2	Codice alfanumerico dell'area sub-comunale attribuito dal comune}
\item{ASC_NOME	Denominazione dell'area sub-comunale}
\item{ASC_TIPO	Tipologia di area-sub-comunale}
\item{CIRC_COD	Codice della circoscrizione elettorale della Camera dei deputati}
\item{CIRC_DEN	Denominazione della circoscrizione elettorale della Camera dei deputati}
\item{CU20_COD	Codice del collegio elettorale uninominale della Camera dei deputati}
\item{CP20_COD	Codice del collegio elettoraleplurinominale della Camera dei deputati}
\item{SU20_COD	Codice del collegio elettorale uninominale del Senato della Repubblica}
\item{SP20_COD	Codice del collegio elettorale plurinominale del Senato della Repubblica}
\item{CU20_DEN	Denominazione del collegio elettorale uninominale della Camera dei deputati}
\item{CP20_DEN	Denominazione del collegio elettorale plurinominale della Camera dei deputati}
\item{SU20_DEN	Denominazione del collegio elettorale uninominale del Senato della Repubblica}
\item{SP20_DEN	Denominazione del collegio elettorale plurinominale del Senato della Repubblica}
\item{CU20_C1	Sigla del collegio elettorale uninominale della Camera dei deputati}
\item{CP20_C1	Sigla del collegio elettorale plurinominale della Camera dei deputati}
\item{SU20_C1	Sigla del collegio elettorale uninominale del Senato della Repubblica}
\item{SP20_C1	Sigla del collegio elettorale plurinominale del Senato della Repubblica}
}
}
\examples{
ll_set_folder(fs::path(fs::path_home_r(), "R"))
ll_get_electoral_districts_it()
ll_get_electoral_districts_it(name = "Lombardia 2")
ll_get_electoral_districts_it() \%>\% ggplot2::ggplot() + ggplot2::geom_sf() + ggplot2::labs(title = "Circoscrizioni Camera")
ll_get_electoral_districts_it(level = "SENATO_CollegiUNINOMINALI_2020") \%>\% ggplot2::ggplot() + ggplot2::geom_sf() + ggplot2::labs(title = "Collegi uninominali - Senato")
}
