% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_nuts.R
\name{ll_get_gadm}
\alias{ll_get_gadm}
\title{Get administrative boundaries}
\usage{
ll_get_gadm(geo, level = 0, version = "4.0")
}
\arguments{
\item{geo}{Three letter country codes. If a two letter country code is given, it will tentatively be converted to a three-letter country code. Check consistency.}

\item{level}{Defaults to 0. Available labels, depending on data availability for the specific country, between 0 and 3.}

\item{version}{Defaults to "4.0". Untested with others.}
}
\value{
An \code{sf} object
}
\description{
Source: https://gadm.org/
}
\examples{
ll_get_gadm(geo = "UKR", level = 2)
}
