% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_lau_eu.R
\name{ll_get_lau_eu}
\alias{ll_get_lau_eu}
\title{Gets local administrative units from Eurostat's website}
\usage{
ll_get_lau_eu(name = NULL, year = 2018, silent = FALSE)
}
\arguments{
\item{year}{Year of mapping, defaults to most recent (2018). Availalble values: 2017, 2018}
}
\value{
European LAU in sf format
}
\description{
Source: https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/lau#lau18
}
\examples{

ll_set_folder("~/R/")
ll_get_lau_eu()

}
