% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_lau_eu.R
\name{ll_get_lau_eu}
\alias{ll_get_lau_eu}
\title{Gets local administrative units from Eurostat's website}
\usage{
ll_get_lau_eu(year = 2018)
}
\value{
LAU in sf format
}
\description{
Gets local administrative units from Eurostat's website
}
\examples{
ll_get_lau_eu()
}
