% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_lau_eu.R
\name{ll_get_lau_eu}
\alias{ll_get_lau_eu}
\title{Gets local administrative units from Eurostat's website}
\usage{
ll_get_lau_eu(name = NULL, year = 2019, silent = FALSE, lau_sf = NULL)
}
\arguments{
\item{year}{Year of mapping, defaults to most recent (2019). Availalble values: 2016, 2017, 2018, 2019.}

\item{lau_sf}{sf object, exactly such as the one that would be returned by \code{ll_get_lau_eu()}. Used to speed-up computation when bulk processing.}
}
\value{
European LAU in sf format
}
\description{
Source: https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/lau#lau18
}
\examples{

ll_set_folder("~/R/")
ll_get_lau_eu()

}
