% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_lau_nuts_concordance.R
\name{ll_get_lau_nuts_concordance}
\alias{ll_get_lau_nuts_concordance}
\title{Gets correspondence tables between local administrative units and nuts from Eurostat's website}
\usage{
ll_get_lau_nuts_concordance(lau_year = 2019, nuts_year = 2016, silent = FALSE)
}
\arguments{
\item{lau_year}{Defaults to 2019. See \code{ll_lau_nuts_concordance_links} for details on available combinations.}

\item{nuts_year}{Defaults to 2016. See \code{ll_lau_nuts_concordance_links} for details on available combinations.}

\item{silent}{Defaults to FALSE. If TRUE, hides copyright notice. Useful e.g. when using this in reports or in loops. The copyright notice must still be shown where the final output is used.}
}
\value{
A tibble with a correspondence table.
}
\description{
Source: https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/lau
}
\details{
Warning: due to issues in the original data, nuts may not always correspond to the given year for all countries, e.g. in files with nuts 2016 one may find nuts 2013 for single country, e.g. Italy.
Do check the sources for details and ensure complete matching.
}
\examples{

ll_set_folder("~/R/")
ll_get_lau_nuts_concordance()

\dontrun{
lau_with_nuts_df <- ll_get_lau_eu(year = 2018) \%>\% 
  sf::st_drop_geometry() \%>\% 
  filter(is.na(LAU_NAME)==FALSE) \%>\% 
  dplyr::rename(gisco_id = GISCO_ID) \%>\% 
  dplyr::left_join(y = ll_get_lau_nuts_concordance(lau_year = 2018,
                                                   nuts_year = 2016),
                   by = "gisco_id")
}
}
