% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_nuts_eu.R
\name{ll_get_nuts_eu}
\alias{ll_get_nuts_eu}
\title{Gets NUTS as sf object from Eurostat's website}
\usage{
ll_get_nuts_eu(level = 3, resolution = 60, year = 2016)
}
\arguments{
\item{level}{Defaults to 3, corresponding to nuts3. Available values are: 0, 1, 2, and 3.}

\item{resolution}{Defaults to "60", for 1:60 Million. Available values: are 20, 10, 3, 1 (1 is highest quality available)-}

\item{year}{Defaults to 2016. Available values: 2016, 2013, 2010, 2006, 2001}
}
\value{
NUTS in sf format
}
\description{
Source: https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/nuts#nuts16
}
\examples{
ll_get_lau_eu()
}
