% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_nuts_it.R
\name{ll_get_nuts_it}
\alias{ll_get_nuts_it}
\title{Regions and provinces in Italy (high detail, CC-BY Istat)}
\usage{
ll_get_nuts_it(
  name = NULL,
  level = 2,
  year = 2020,
  resolution = "low",
  silent = FALSE
)
}
\arguments{
\item{level}{Defaults to "2", i.e. regioni. Available: "3" (i.e. province), and "lau", local administrative units.}

\item{year}{Defaults to 2020 (latest available).}
}
\value{

}
\description{
2019 / WGS84 UTM32N
}
\examples{
ll_set_folder("~/R/")
ll_get_nuts_it()
ll_get_nuts_it(name = "Rimini")
}
