% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_population_grid.R
\name{ll_get_population_grid}
\alias{ll_get_population_grid}
\title{Get EU 1km population grid}
\usage{
ll_get_population_grid(year = 2011)
}
\value{
An sf object with the population grid.
}
\description{
Source: https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/population-distribution-demography/geostat
More details: https://ec.europa.eu/eurostat/statistics-explained/index.php/Population_grids
}
