% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_get_population_grid.R
\name{ll_get_population_grid}
\alias{ll_get_population_grid}
\title{Get EU 1km population grid}
\usage{
ll_get_population_grid(
  year = 2011,
  match_sf = NULL,
  match_name = NULL,
  match_country = NULL,
  join = sf::st_intersects,
  silent = FALSE
)
}
\arguments{
\item{year}{Defaults to 2011. Currently, the EU population grid is available only for the year 2011.}

\item{match_sf}{An sf object to me matched with the population grid. If not given, full grid is returned.}

\item{match_name}{A name to be used for local caching. It is the responsibility of the user to keept it consistent. If not given, data are not cached locally.}

\item{match_country}{Defaults to NULL. If given, used to speed up processing.}
}
\value{
An sf object with the population grid.
}
\description{
Source: https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/population-distribution-demography/geostat
More details: https://ec.europa.eu/eurostat/statistics-explained/index.php/Population_grids
}
